@@ -1,91 +1,46 @@
 {
     "name": "kevinboutin",
-    "version": "2.0.3",
+    "version": "2.1.4",
     "description": "A professional view of the career of Kevin Boutin",
     "author": "Kevin Boutin <me@kevinboutin.me>",
     "private": true,
+    "type": "module",
     "license": "GPL-3.0",
     "repository": {
         "type": "git",
         "url": "https://github.com/kevboutin/kevinboutin.git"
     },
     "scripts": {
-        "dev": "webpack-dev-server --inline --progress --config build/webpack.dev.conf.js",
-        "start": "npm run dev",
-        "unit": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js --single-run",
-        "test": "npm run unit",
-        "lint": "eslint --ext .js,.vue src test/unit",
-        "build": "node build/build.js"
+        "dev": "vite",
+        "build": "vite build",
+        "preview": "vite preview",
+        "test": "npm run test:unit",
+        "test:unit": "vitest",
+        "lint": "eslint . --fix",
+        "format": "prettier --write src/"
     },
     "dependencies": {
-        "echarts": "^4.2.1",
-        "vue": "^2.6.10",
-        "vue-router": "^3.0.1"
+        "echarts": "^5.6.0",
+        "vue": "^3.5.16",
+        "vue-router": "^4.5.1"
     },
     "devDependencies": {
-        "autoprefixer": "^7.1.2",
-        "babel-core": "^6.22.1",
-        "babel-eslint": "^8.2.1",
-        "babel-helper-vue-jsx-merge-props": "^2.0.3",
-        "babel-loader": "^7.1.1",
-        "babel-plugin-istanbul": "^4.1.1",
-        "babel-plugin-syntax-jsx": "^6.18.0",
-        "babel-plugin-transform-runtime": "^6.22.0",
-        "babel-plugin-transform-vue-jsx": "^3.5.0",
-        "babel-preset-env": "^1.3.2",
-        "babel-preset-stage-2": "^6.22.0",
-        "chai": "^4.2.0",
-        "chalk": "^2.4.2",
-        "copy-webpack-plugin": "^4.6.0",
-        "cross-env": "^5.0.1",
-        "css-loader": "^0.28.11",
-        "eslint": "^4.15.0",
-        "eslint-config-google": "^0.9.1",
-        "eslint-friendly-formatter": "^3.0.0",
-        "eslint-loader": "^1.7.1",
-        "eslint-plugin-html": "^4.0.6",
-        "eslint-plugin-vue": "^4.0.0",
-        "extract-text-webpack-plugin": "^3.0.0",
-        "file-loader": "^1.1.4",
-        "friendly-errors-webpack-plugin": "^1.6.1",
-        "html-webpack-plugin": "^2.30.1",
-        "inject-loader": "^3.0.0",
-        "karma": "^2.0.5",
-        "karma-coverage": "^1.1.1",
-        "karma-mocha": "^1.3.0",
-        "karma-phantomjs-launcher": "^1.0.2",
-        "karma-phantomjs-shim": "^1.4.0",
-        "karma-sinon-chai": "^1.3.1",
-        "karma-sourcemap-loader": "^0.3.7",
-        "karma-spec-reporter": "0.0.31",
-        "karma-webpack": "^2.0.2",
-        "mocha": "^5.2.0",
-        "node-notifier": "^5.4.0",
-        "optimize-css-assets-webpack-plugin": "^3.2.0",
-        "ora": "^1.2.0",
-        "phantomjs-prebuilt": "^2.1.14",
-        "portfinder": "^1.0.20",
-        "postcss-import": "^11.0.0",
-        "postcss-loader": "^2.0.8",
-        "postcss-url": "^7.2.1",
-        "rimraf": "^2.6.3",
-        "semver": "^5.7.0",
-        "shelljs": "^0.7.6",
-        "sinon": "^4.0.0",
-        "sinon-chai": "^2.8.0",
-        "uglifyjs-webpack-plugin": "^1.3.0",
-        "url-loader": "^1.1.2",
-        "vue-loader": "^13.7.3",
-        "vue-style-loader": "^3.1.2",
-        "vue-template-compiler": "^2.6.10",
-        "webpack": "^3.6.0",
-        "webpack-bundle-analyzer": "^3.3.2",
-        "webpack-dev-server": "^2.9.1",
-        "webpack-merge": "^4.1.0"
+        "@eslint/js": "^9.28.0",
+        "@vitejs/plugin-vue": "^5.2.4",
+        "@vitest/eslint-plugin": "1.2.1",
+        "@vue/eslint-config-prettier": "^10.2.0",
+        "@vue/test-utils": "^2.4.6",
+        "eslint": "^9.28.0",
+        "eslint-plugin-vue": "^10.2.0",
+        "jsdom": "^26.1.0",
+        "prettier": "^3.5.3",
+        "vite": "^6.3.5",
+        "vite-plugin-vue-devtools": "^7.7.6",
+        "vitest": "^3.2.3"
     },
     "engines": {
-        "node": ">= 6.0.0",
-        "npm": ">= 3.0.0"
+        "node": ">= 20.0.0",
+        "npm": ">= 10.0.0"
     },
     "browserslist": [
         "> 1%",
