@@ -1,15 +1,10 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue';
-import App from './App';
-import router from './router';
+import { createApp } from "vue";
 
-Vue.config.productionTip = false;
+import App from "./App.vue";
+import router from "./router";
 
-/* eslint-disable no-new */
-new Vue({
-    el: '#app',
-    router,
-    components: {App},
-    template: '<App/>'
-});
+const app = createApp(App);
+
+app.use(router);
+
+app.mount("#app");
