@@ -1,24 +1,7 @@
-<template>
-    <div id="app">
-        <router-view/>
-    </div>
-</template>
-
-<script>
-export default {
-  name: 'App'
-};
+<script setup>
+import { RouterView } from 'vue-router'
 </script>
 
-<style>
-    body {
-        margin: 0;
-        background: url('./assets/noise.png');
-    }
-    #app {
-        font: 16px/26px 'Avenir', Helvetica, Helvetica Neue, Arial, sans-serif;
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-        color: #2c3e50;
-    }
-</style>
+<template>
+    <RouterView />
+</template>
