@@ -1,27 +1,23 @@
-import Vue from 'vue';
-import Router from 'vue-router';
-import Home from '@/components/Home';
-import Auto from '@/components/Auto';
-import Cooking from '@/components/Cooking';
+import { createRouter, createWebHistory } from "vue-router";
+import HomeView from "../views/HomeView.vue";
 
-Vue.use(Router);
-
-export default new Router({
-  routes: [
-    {
-      path: '/',
-      name: 'Home',
-      component: Home
-    },
-    {
-      path: '/recommends/auto',
-      name: 'Auto',
-      component: Auto
-    },
-    {
-      path: '/recommends/cooking',
-      name: 'Cooking',
-      component: Cooking
-    }
-  ]
+const router = createRouter({
+    history: createWebHistory(import.meta.env.BASE_URL),
+    routes: [
+        {
+            path: "/",
+            name: "home",
+            component: HomeView
+        },
+        {
+            path: "/about",
+            name: "about",
+            // route level code-splitting
+            // this generates a separate chunk (About.[hash].js) for this route
+            // which is lazy-loaded when the route is visited.
+            component: () => import("../views/AboutView.vue")
+        }
+    ]
 });
+
+export default router;
